requirements.txt
streamlit
langchain  # <--- Add this line
qdrant-client
langchain-groq
langchain-community
fastembed
pypdf
python-dotenv
pydantic>=2.5.0